requirements.txt
# Существующие зависимости парсера
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
pathlib==1.0.1
lxml==4.9.3

# Telegram бот
python-telegram-bot==20.8

# OpenAI API
openai==1.12.0

# Обработка текста и эмбеддинги (optional для RAG)
sentence-transformers==2.2.2
numpy<2.0.0
scikit-learn==1.3.0

# База данных (sqlite3 встроен в Python)

# Конфигурация
python-dotenv==1.0.0

# Логирование и утилиты
python-json-logger==2.0.7